add 横スクロール遷移とスワイプ下遷移
