+++ AKScheduler/SchedulerViewController.swift
@@ -12,8 +12,7 @@ class SchedulerViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
+        setupGestureRecognizer()
     }
 
     override func didReceiveMemoryWarning() {
@@ -21,6 +20,21 @@ class SchedulerViewController: UIViewController {
         // Dispose of any resources that can be recreated.
     }
     
+    ///スワイプ処理の初期化
+    func setupGestureRecognizer() {
+        //スワイプを判断するための変数を定義
+        let leftSwipeRecognizer  = UISwipeGestureRecognizer(target: self, action: #selector(leftSwiped(_:)))
+        //スワイプの方向を判断
+        leftSwipeRecognizer.direction  = .left
+        //viewに追加
+        self.view.addGestureRecognizer(leftSwipeRecognizer)
+    }
+    //MARK: - UISwipeGestureを用いて画面遷移
+    //左へスワイプしたときに呼ばれる関数
+    @objc func leftSwiped(_ sender: Any) {
+        print("leftSwipe!")
+        dismiss(animated: false, completion: nil)
+    }
 
     /*
     // MARK: - Navigation
